requirements.txt
alembic==1.0.0
aniso8601==3.0.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
coverage==4.5.1
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
funcsigs==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mock==2.0.0
nose2==0.7.4
pbr==4.2.0
psycopg2==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
Werkzeug==0.14.1
